feat: add wizard-like file configuration ui
