+++ frontend/src/features/coordinate_upload/stepper_pages/ValidationScreen.tsx
@@ -1,0 +1,5 @@
+import { Box } from "@mui/material";
+
+export default function CoordinateValidationScreen() {
+  return <Box>Hello World</Box>;
+}
